requirements.txt
numpy==1.24.3
Pillow==10.0.0
flask==3.0.0